Add database tests for user model
There is a new seed file for users that creates one test user and a bunch of randomized users; the tests check creation, retrieval, update, and delete functionality. IMPORTANT: TestCase.php now changes the database to use sqllite (just in memory). This is to make the data access faster for the unit tests. If you aren't set up for sqllite, you may have to install it, or comment out this line (but your tests may be slower).